refactor: move intractions with card to separate files
