+++ src/features/catalog/components/ProductList/ProductList.tsx
@@ -7,13 +7,9 @@ import { useCategory } from '../../../../contexts/CategoryContext.tsx';
 import type { SortValues } from '../../types/sort.ts';
 import formatDataForSticker from '../../../../utils/formatDataForSticker/formatDataForSticker.ts';
 import { Typography } from '@mui/material';
-import { useAddLineItemMutation, useGetMyActiveCartQuery } from '../../../../api/cartApi.ts';
-import type { Cart } from '../../../../types/cartApi.ts';
-import { useSnackbar } from 'notistack';
-
-function isProductInCart(productId: string, cart?: Cart): boolean {
-  return cart ? cart.lineItems.some((item) => item.productId === productId) : false;
-}
+import { useGetMyActiveCartQuery } from '../../../../api/cartApi.ts';
+import { isProductInCart } from '../../../../utils/isProductInCart.ts';
+import { useAddToCart } from '../../../../hooks/useAddToCart.ts';
 
 export default function ProductList({
   sortValue,
@@ -32,8 +28,6 @@ export default function ProductList({
     selectedCategory,
   } = useCategory();
 
-  const { enqueueSnackbar } = useSnackbar();
-
   const {
     data,
     isLoading: isProductLoading,
@@ -46,20 +40,9 @@ export default function ProductList({
     selectedPriceRange: priceRange,
   });
 
-  const addItem = useAddLineItemMutation()[0];
   const { data: cart } = useGetMyActiveCartQuery();
 
-  const handleAddToCart = async (id: string) => {
-    if (!cart) return;
-    const { data: updatedCart } = await addItem({
-      cartId: cart.id,
-      version: cart.version,
-      draft: { productId: id },
-    });
-    if (isProductInCart(id, updatedCart)) {
-      enqueueSnackbar('Product is added to your cart', { variant: 'success' });
-    }
-  };
+  const handleAddToCart = useAddToCart(cart);
 
   if (isCategoryLoading || isProductLoading) {
     return <CircularProgress size="3rem" />;
